// web/src/features/dashboard/types.ts
export interface UserChartsFilters {
  timeGranularity: TimeGranularity
  range: { start: Date; end: Date }
  topUserLimit: number
}